refactor: 파편화되어있던 slice 병합 및 로직 리팩토링 (#ATR-390)
