--- apps/service/entities/newsletter-previous-article-item/model/type/index.ts
@@ -1,1 +1,0 @@
-export type { default as NewsletterPreviousArticleData } from './newsletterPreviousArticleData'
