--- apps/service/entities/newsletter-previous-article-item/model/type/newsletterPreviousArticleData.ts
@@ -1,10 +1,0 @@
-export default interface NewsletterPreviousArticleData {
-  id: number
-  title: string
-  thumbnailUrl: string
-  contentUrl: string
-  contentSummary: string
-  readingTime: number
-  receivedAt: string
-  newsletterName: string
-}
