--- apps/service/entities/newsletter-previous-article-item/ui/index.ts
@@ -1,1 +1,0 @@
-export { default as NewsletterPreviousArticleItem } from './NewsletterPreviousArticleItem'
