--- apps/service/entities/related-newsletter-item/index.ts
@@ -1,2 +1,0 @@
-export * from './model'
-export * from './ui'
