--- apps/service/entities/related-newsletter-item/model/index.ts
@@ -1,1 +1,0 @@
-export * from './type'
