--- apps/service/entities/related-newsletter-item/model/type/index.ts
@@ -1,1 +1,0 @@
-export type { default as RelatedNewsletter } from './relatedNewsletter'
