--- apps/service/entities/related-newsletter-item/model/type/relatedNewsletter.ts
@@ -1,6 +1,0 @@
-export default interface RelatedNewsletter {
-  id: string
-  thumbnailUrl: string
-  name: string
-  description: string
-}
