--- apps/service/entities/related-newsletter-item/ui/index.ts
@@ -1,1 +1,0 @@
-export { default as RelatedNewsletterItem } from './RelatedNewsletterItem'
