--- apps/service/features/newsletter-previous-articles/api/index.ts
@@ -1,1 +1,0 @@
-export { default as fetchNewsletterPreviousArticles } from './fetchNewsletterPreviousArticles'
