--- apps/service/features/newsletter-previous-articles/index.ts
@@ -1,4 +1,0 @@
-export * from './ui'
-export * from './api'
-export * from './lib'
-export * from './model'
