--- apps/service/features/newsletter-previous-articles/lib/hook/index.ts
@@ -1,1 +1,0 @@
-export { default as useNewsletterPreviousArticles } from './useNewsletterPreviousArticles'
