--- apps/service/features/newsletter-previous-articles/model/index.ts
@@ -1,1 +1,0 @@
-export * from './type'
