--- apps/service/features/newsletter-previous-articles/model/type/NewsletterPreviousArticlesResponse.ts
@@ -1,7 +1,0 @@
-import { NewsletterPreviousArticleData } from '@/entities/newsletter-previous-article-item'
-
-export interface NewsletterPreviousArticlesResponse {
-  status: number
-  message: string
-  data: NewsletterPreviousArticleData[]
-}
