--- apps/service/features/newsletter-previous-articles/model/type/index.ts
@@ -1,1 +1,0 @@
-export * from './NewsletterPreviousArticlesResponse'
