--- apps/service/features/newsletter-previous-articles/ui/index.ts
@@ -1,1 +1,0 @@
-export { default as NewsletterPreviousArticles } from './NewsletterPreviousArticles'
