--- apps/service/features/newsletter-profile/api/index.ts
@@ -1,2 +1,0 @@
-export { default as fetchNewsletterProfile } from './fetchNewsletterProfile'
-export { default as subscribeNewsletter } from './subscribeNewsletter'
