--- apps/service/features/newsletter-profile/index.ts
@@ -1,4 +1,0 @@
-export * from './api'
-export * from './model'
-export * from './ui'
-export * from './lib'
