--- apps/service/features/newsletter-profile/lib/hook/index.ts
@@ -1,1 +1,0 @@
-export { default as useNewsletterProfile } from './useNewsletterProfile'
