--- apps/service/features/newsletter-profile/lib/index.ts
@@ -1,1 +1,0 @@
-export * from './hook'
