--- apps/service/features/newsletter-profile/ui/index.ts
@@ -1,1 +1,0 @@
-export { default as NewsletterProfile } from './NewsletterProfile'
