--- apps/service/features/related-newsletters/api/index.ts
@@ -1,1 +1,0 @@
-export { default as fetchRelatedNewsletters } from './fetchRelatedNewsletters'
