--- apps/service/features/related-newsletters/index.ts
@@ -1,4 +1,0 @@
-export * from './api'
-export * from './lib'
-export * from './model'
-export * from './ui'
