--- apps/service/features/related-newsletters/lib/hook/index.ts
@@ -1,1 +1,0 @@
-export { default as useRelatedNewsletters } from './useRelatedNewsletters'
