--- apps/service/features/related-newsletters/lib/index.ts
@@ -1,1 +1,0 @@
-export * from './hook'
