--- apps/service/features/related-newsletters/model/index.ts
@@ -1,1 +1,0 @@
-export * from './type'
