--- apps/service/features/related-newsletters/model/type/index.ts
@@ -1,1 +1,0 @@
-export type { default as RelatedNewslettersResponse } from './relatedNewslettersResponse'
