--- apps/service/features/related-newsletters/model/type/relatedNewslettersResponse.ts
@@ -1,7 +1,0 @@
-import { RelatedNewsletter } from '@/entities/related-newsletter-item'
-
-export default interface RelatedNewslettersResponse {
-  status: string
-  message: string
-  data: RelatedNewsletter[]
-}
