--- apps/service/features/related-newsletters/ui/index.ts
@@ -1,1 +1,0 @@
-export { default as RelatedNewsletters } from './RelatedNewsletters'
