+++ apps/service/widgets/newsletter-introduce/api/index.ts
@@ -1,0 +1,4 @@
+export { default as fetchNewsletterProfile } from './fetchNewsletterProfile'
+export { default as subscribeNewsletter } from './subscribeNewsletter'
+export { default as fetchNewsletterPreviousArticles } from './fetchNewsletterPreviousArticles'
+export { default as fetchRelatedNewsletters } from './fetchRelatedNewsletters'
