+++ apps/service/widgets/newsletter-introduce/index.ts
@@ -1,1 +1,4 @@
+export * from './api'
+export * from './lib'
+export * from './model'
 export * from './ui'
