+++ apps/service/widgets/newsletter-introduce/lib/hook/index.ts
@@ -1,0 +1,3 @@
+export { default as useNewsletterProfile } from './useNewsletterProfile'
+export { default as useNewsletterPreviousArticles } from './useNewsletterPreviousArticles'
+export { default as useRelatedNewsletters } from './useRelatedNewsletters'
