+++ apps/service/widgets/newsletter-introduce/model/index.ts
