+++ apps/service/widgets/newsletter-introduce/model/type/index.ts
@@ -1,0 +1,3 @@
+export * from './newsletterPreviousArticlesResponse'
+export * from './newsletterProfileResponse'
+export * from './relatedNewslettersResponse'
