+++ apps/service/widgets/newsletter-introduce/model/type/newsletterPreviousArticlesResponse.ts
@@ -1,0 +1,16 @@
+export interface NewsletterPreviousArticlesResponse {
+  status: number
+  message: string
+  data: NewsletterPreviousArticleData[]
+}
+
+export interface NewsletterPreviousArticleData {
+  id: number
+  title: string
+  thumbnailUrl: string
+  contentUrl: string
+  contentSummary: string
+  readingTime: number
+  receivedAt: string
+  newsletterName: string
+}
