+++ apps/service/widgets/newsletter-introduce/model/type/newsletterProfileResponse.ts
@@ -1,5 +1,11 @@
 import { NewsletterCategory } from '@/shared/type'
 
+export interface NewsletterProfileResponse {
+  status: string
+  message: string
+  data: NewsletterProfileData
+}
+
 export interface NewsletterProfileData {
   name: string
   description: string
@@ -8,10 +14,4 @@ export interface NewsletterProfileData {
   mainLink: string
   subscribeLink: string
   thumbnailUrl: string
-}
-
-export interface NewsletterProfileResponse {
-  status: string
-  message: string
-  data: NewsletterProfileData
 }
