+++ apps/service/widgets/newsletter-introduce/model/type/relatedNewslettersResponse.ts
@@ -1,0 +1,12 @@
+export interface RelatedNewslettersResponse {
+  status: string
+  message: string
+  data: RelatedNewsletter[]
+}
+
+export interface RelatedNewsletter {
+  id: string
+  thumbnailUrl: string
+  name: string
+  description: string
+}
