+++ apps/service/widgets/newsletter-introduce/ui/RelatedNewsletterItem.tsx
